Fixed the DeckSearchBar, DeckScrollList now contains ScrollListEntry widgets, instead of strings
DeckSearchBar now displays Search... when out of focus, searches among
the deck titles as the user types, and the results become visible on the
scroll list.
The DeckScrollList now contains DeckScrollListEntry widgets instead of
QStrings, so they are more customizable. Added an Edit button for
DeckScrollListEntry widgets, that sends a signal, that gets forwarded
all the way to CentralWidget, where it is rerouted to DeckEditor(will
be, once DeckEditorTab is developed)
DeckScrollList now uses DeckScrollArea instead of QScrollArea, this is
for sizeHint() customization.